--- Assets/Animation/AirGuardAnim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a2791537060d7e842b2c3d1daf32cee4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
