--- Assets/Animation/AirGuardAnim/Blockfinal.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3127eba3c7e37cff8bbd88df35bc18d4
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
