--- Assets/Animation/AirGuardAnim/Feintfinal.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3bffa62ab997163338bbbffc90ab3461
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
