--- Assets/Animation/AirGuardAnim/Guard Controlleridk.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fc3555dfa66a5d84c8d66bb882366802
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
