--- Assets/Animation/AirGuardAnim/None.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: dc4630cbfea5f864882717735ea09d75
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
