--- Assets/Animation/AirGuardAnim/heavyattackhopefullyfinal.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5e9bc2493acff52428904616fcb6c089
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
