--- Assets/Animation/AirGuardAnim/quickattackfinal.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7125090931f02195e829817835ce3652
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
