--- Assets/Animation/WizardAnim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 525201fd0f213584ba6c8e63bb618800
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
