+++ Assets/Art/Battle/Enemies/AirGeneral.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 21b929b218c976b4ca10c0e2e69647a7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
