+++ Assets/Art/Battle/Enemies/AirGuard.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a420b4214d9ccc948808a454782b6f80
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
