--- Assets/Earth.json.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: bef2dc1ab3a0cbe4999d46720b567191
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
