+++ Assets/Scripts/Combat/Spell/FireBall.cs
@@ -13,7 +13,7 @@ public class FireBall : MonoBehaviour
 
     private void OnTriggerEnter2D(Collider2D collision)
     {
-        if (collision.tag == "Enemy")
+        if (collision.tag == "Enemy" && collision.isTrigger)
         {
             collision.GetComponent<EnemyHealth>().TakeDamage(25.0f);
             Instantiate(FireWall, transform.position, Quaternion.identity);
