+++ Assets/Scripts/Combat/Spell/FireWall.cs
@@ -17,7 +17,7 @@ public class FireWall : MonoBehaviour
 
     private void OnTriggerStay2D(Collider2D collider)
     {
-        if (collider.tag == "Enemy")
+        if (collider.tag == "Enemy" && collider.isTrigger)
         {
             collider.gameObject.GetComponent<EnemyHealth>().TakeDamage(10f * Time.deltaTime);
         }
