--- Assets/Scripts/SealBoss.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 63ee8fe6d2eec4d41835e3dcf909d576
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
