+++ UserSettings/EditorUserSettings.asset
@@ -24,10 +24,10 @@ EditorUserSettings:
       value: 07000100520208590b0b5b7248705b4447151e7979702561797f4931bae2313b
       flags: 0
     RecentlyUsedSceneGuid-6:
-      value: 500404525402505a0e0f0f7340210944464f1b7e28297e602c284564e3e1303a
+      value: 55535055515659595f575c2749200744474e4f2e7a782765747a4b62e6b9616c
       flags: 0
     RecentlyUsedSceneGuid-7:
-      value: 55535055515659595f575c2749200744474e4f2e7a782765747a4b62e6b9616c
+      value: 500404525402505a0e0f0f7340210944464f1b7e28297e602c284564e3e1303a
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
